update manage server page
