+++ go.mod
@@ -2,6 +2,8 @@ module github.com/spiffe/tornjak
 
 go 1.22
 
+toolchain go1.22.2
+
 require (
 	github.com/MicahParks/keyfunc/v2 v2.1.0
 	github.com/cenkalti/backoff/v4 v4.2.0
